Rework mapy
- Tileset i mapa go wykorzystująca
- Poprawa zaczętego systemu announcera fal
- System dopasowywania miejsca kliknięcia do siatki, pomijający przy tym UI

+++ Tower Defense Mobile/Assets/Enemies/Enemy.cs
@@ -21,7 +21,7 @@ public abstract class Enemy : MonoBehaviour{
         return new[] { (float)currentTargetIndex, distanceFromNextWaypoint };
     }
 
-    public void TakeDamage(int dmg) {
+    public void TakeDamage(float dmg) {
         health -= dmg;
         if (health <= 0) {
             Die();
